refactor: improve `slotviz` section's component structure and namings
