+++ frontend/.vscode/settings.json
@@ -13,6 +13,7 @@
         "DashboardChartSummary",
         "DashboardChartSummaryFilter",
         "DashboardGroupManagementModal",
+        "DashboardSlotViz",
         "DashboardTableSummaryDetails",
         "DashboardTableSummaryValue",
         "DashboardTableValidators",
@@ -33,7 +34,6 @@
         "NotificationsNetworkTable",
         "NotificationsOverview",
         "NotificationsTableEmpty",
-        "SlotVizViewer",
         "a11y",
         "checkout",
         "ci",
