+++ frontend/components/dashboard/slot-viz/DashboardSlotVizDutyIcon.vue
